# rezara-app/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Confirmation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AuthUser {
  id: string;
  email: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string | null;
  role: AuthUserRole;
}

export interface Business {
  id: number;
  userId: string;
  name: string;
  logo?: string | null;
  phone?: string | null;
  address?: string | null;
  description?: string | null;
  createdAt: string;
}

export interface UpsertBusinessBody {
  name: string;
  logo?: string | null;
  phone?: string | null;
  address?: string | null;
  description?: string | null;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  expired: "expired",
} as const;

export interface Reservation {
  id: number;
  businessId: number;
  linkId: string;
  customerName: string;
  customerPhone?: string | null;
  date: string;
  time: string;
  guests: number;
  depositAmount: number;
  notes?: string | null;
  status: ReservationStatus;
  createdAt: string;
  updatedAt: string;
}

export type ReservationWithBusiness = Reservation & {
  business: Business;
};

export interface CreateReservationBody {
  /** Optional client-generated link ID (10-char alphanumeric). Server generates one if omitted. */
  linkId?: string;
  customerName: string;
  customerPhone?: string | null;
  date: string;
  time: string;
  guests: number;
  depositAmount: number;
  notes?: string | null;
}

export type UpdateReservationBodyStatus =
  (typeof UpdateReservationBodyStatus)[keyof typeof UpdateReservationBodyStatus];

export const UpdateReservationBodyStatus = {
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  expired: "expired",
} as const;

export interface UpdateReservationBody {
  customerName?: string;
  customerPhone?: string | null;
  date?: string;
  time?: string;
  guests?: number;
  depositAmount?: number;
  notes?: string | null;
  status?: UpdateReservationBodyStatus;
}

export type PaymentPaymentStatus =
  (typeof PaymentPaymentStatus)[keyof typeof PaymentPaymentStatus];

export const PaymentPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  refunded: "refunded",
} as const;

export interface Payment {
  id: number;
  reservationId: number;
  businessId: number;
  amount: number;
  currency: string;
  paymentStatus: PaymentPaymentStatus;
  paypalOrderId?: string | null;
  createdAt: string;
  customerName?: string | null;
  reservationDate?: string | null;
}

export interface PayPalConfig {
  clientId: string;
  mode: string;
}

export interface CreatePayPalOrderBody {
  linkId: string;
}

export interface PayPalOrderResponse {
  orderId: string;
}

export type ChangeReservationStatusBodyStatus =
  (typeof ChangeReservationStatusBodyStatus)[keyof typeof ChangeReservationStatusBodyStatus];

export const ChangeReservationStatusBodyStatus = {
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  expired: "expired",
} as const;

export interface ChangeReservationStatusBody {
  status: ChangeReservationStatusBodyStatus;
}

export interface CapturePayPalOrderBody {
  orderId: string;
  linkId: string;
}

export interface CapacitySlot {
  id: number;
  businessId: number;
  date: string;
  startTime?: string | null;
  durationHours?: number | null;
  maxCapacity: number;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertCapacitySlotBody {
  id?: number;
  date: string;
  startTime?: string | null;
  durationHours?: number | null;
  maxCapacity: number;
}

export interface NotificationItem {
  id: number;
  reservationId: number;
  customerName: string;
  amount: number;
  paidAt: string;
}

export interface DashboardStats {
  totalReservations: number;
  pendingPayment: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  totalRevenue: number;
  recentReservations: Reservation[];
}

export type GetReservationsParams = {
  status?: GetReservationsStatus;
  date?: string;
};

export type GetReservationsStatus =
  (typeof GetReservationsStatus)[keyof typeof GetReservationsStatus];

export const GetReservationsStatus = {
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  expired: "expired",
} as const;

export type GetCapacitySlotsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};

export type DeleteCapacitySlot200 = {
  success: boolean;
};