# rezara-app/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Booking Confirmation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/me:
    get:
      operationId: getMyBusiness
      tags: [businesses]
      summary: Get current user's business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: No business found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertBusiness
      tags: [businesses]
      summary: Create or update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBusinessBody"
      responses:
        "200":
          description: Business saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /payments/paypal/create-order:
    post:
      operationId: createPayPalOrder
      tags: [payments]
      summary: Create a PayPal order for a reservation deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayPalOrderBody"
      responses:
        "200":
          description: PayPal order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayPalOrderResponse"
  /payments/paypal/capture-order:
    post:
      operationId: capturePayPalOrder
      tags: [payments]
      summary: Capture an approved PayPal order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapturePayPalOrderBody"
      responses:
        "200":
          description: Payment captured successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /capacity:
    get:
      operationId: getCapacitySlots
      tags: [capacity]
      summary: Get capacity slots for current business (optionally filtered by date)
      parameters:
        - in: query
          name: date
          schema:
            type: string
          description: Filter by date (YYYY-MM-DD)
      responses:
        "200":
          description: List of capacity slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CapacitySlot"
    post:
      operationId: upsertCapacitySlot
      tags: [capacity]
      summary: Create or update a capacity slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCapacitySlotBody"
      responses:
        "200":
          description: Created or updated capacity slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapacitySlot"
  /capacity/{id}:
    delete:
      operationId: deleteCapacitySlot
      tags: [capacity]
      summary: Delete a capacity slot
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
      required:
        - id
        - email
        - role
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        logo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
    UpsertBusinessBody:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - name
...
    CreatePayPalOrderBody:
      type: object
      properties:
        linkId:
          type: string
      required:
        - linkId
    PayPalOrderResponse:
      type: object
      properties:
        orderId:
          type: string
      required:
        - orderId
...
    CapturePayPalOrderBody:
      type: object
      properties:
        orderId:
          type: string
        linkId:
          type: string
      required:
        - orderId
        - linkId
    CapacitySlot:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        date:
          type: string
        startTime:
          type: string
          nullable: true
        durationHours:
          type: integer
          nullable: true
        maxCapacity:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - date
        - maxCapacity
        - createdAt
        - updatedAt
    UpsertCapacitySlotBody:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        startTime:
          type: string
          nullable: true
        durationHours:
          type: integer
          nullable: true
        maxCapacity:
          type: integer
      required:
        - date
        - maxCapacity
...