# rezara-app/lib/api-zod/src/generated/types/capturePayPalOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Confirmation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CapturePayPalOrderBody {
  orderId: string;
  linkId: string;
}