# rezara-app/lib/api-zod/src/generated/types/upsertCapacitySlotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Confirmation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertCapacitySlotBody {
  id?: number;
  date: string;
  startTime?: string | null;
  durationHours?: number | null;
  maxCapacity: number;
}